1.change video
2.mute
3.full screen

+++ backend/src/main/java/com/quantil/webrtc/signal/bean/MeetingRoomClientRes.java
@@ -1,12 +1,9 @@
 package com.quantil.webrtc.signal.bean;
 
-import com.quantil.webrtc.api.v1.meeting.bean.RtcMeetingItem;
 import lombok.Data;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 /**
  * 应答给客户端的数据
@@ -17,7 +14,7 @@ import java.util.Map;
 @Data
 public class MeetingRoomClientRes {
 
-    private Integer roomId;
+    private Long roomId;
 
     private Integer maxMembers; // 允许最大会议人数
     private String speaker;     // 主讲人
